refactor: serialization utils as standalone file
