+++ src/Util.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "Utils/Serialize.h"
+
 /**
  @def GET_INSTANCE_MEMBER
  @brief Set variable in current namespace based on instance member from GetRuntimeData or GetVRRuntimeData.
@@ -140,18 +142,3 @@ namespace Util
 	bool IsDynamicResolution();
 
 }
-
-namespace nlohmann
-{
-	void to_json(json& j, const float2& v);
-	void from_json(const json& j, float2& v);
-	void to_json(json& j, const float3& v);
-	void from_json(const json& j, float3& v);
-	void to_json(json& j, const float4& v);
-	void from_json(const json& j, float4& v);
-
-	void to_json(json& j, const ImVec2& v);
-	void from_json(const json& j, ImVec2& v);
-	void to_json(json& j, const ImVec4& v);
-	void from_json(const json& j, ImVec4& v);
-};
